+++ src/main/java/com/github/victools/jsonschema/module/javax/validation/JavaxValidationModule.java
@@ -24,8 +24,12 @@ import com.github.victools.jsonschema.generator.SchemaGeneratorConfigBuilder;
 import com.github.victools.jsonschema.generator.SchemaGeneratorConfigPart;
 import java.lang.annotation.Annotation;
 import java.math.BigDecimal;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
 import javax.validation.constraints.DecimalMax;
 import javax.validation.constraints.DecimalMin;
+import javax.validation.constraints.Email;
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.Negative;
@@ -34,6 +38,7 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Null;
+import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Positive;
 import javax.validation.constraints.PositiveOrZero;
 import javax.validation.constraints.Size;
@@ -44,10 +49,22 @@ import javax.validation.constraints.Size;
  * <li>Determine whether a member is not nullable, base assumption being that all fields and method return values are nullable if not annotated.</li>
  * <li>Populate "minItems" and "maxItems" for containers (i.e. arrays and collections).</li>
  * <li>Populate "minLength" and "maxLength" for strings.</li>
+ * <li>Optionally: populate "pattern" for strings.</li>
  * <li>Populate "minimum"/"exclusiveMinimum" and "maximum"/"exclusiveMaximum" for numbers.</li>
  * </ul>
  */
 public class JavaxValidationModule implements Module {
+
+    private final List<JavaxValidationOption> options;
+
+    /**
+     * Constructor.
+     *
+     * @param options features to enable
+     */
+    public JavaxValidationModule(JavaxValidationOption... options) {
+        this.options = options == null ? Collections.emptyList() : Arrays.asList(options);
+    }
 
     @Override
     public void applyToConfigBuilder(SchemaGeneratorConfigBuilder builder) {
@@ -56,7 +73,7 @@ public class JavaxValidationModule implements Module {
     }
 
     /**
-     * Apply the various annotation-based resolvers for the given configuration part (this expected to be executed for both fields and methods).
+     * Apply the various annotation-based resolvers for the given configuration part (this is expected to be executed for both fields and methods).
      *
      * @param configPart config builder part to add configurations to
      */
@@ -70,6 +87,10 @@ public class JavaxValidationModule implements Module {
         configPart.withNumberExclusiveMinimumResolver(this::resolveNumberExclusiveMinimum);
         configPart.withNumberInclusiveMaximumResolver(this::resolveNumberInclusiveMaximum);
         configPart.withNumberExclusiveMaximumResolver(this::resolveNumberExclusiveMaximum);
+
+        if (this.options.contains(JavaxValidationOption.INCLUDE_PATTERN_EXPRESSIONS)) {
+            configPart.withStringPatternResolver(this::resolveStringPattern);
+        }
     }
 
     /**
@@ -197,6 +218,29 @@ public class JavaxValidationModule implements Module {
             if (sizeAnnotation != null && sizeAnnotation.max() < 2147483647) {
                 // maximum length below the default 2147483647 was specified
                 return sizeAnnotation.max();
+            }
+        }
+        return null;
+    }
+
+    /**
+     * Determine a given text type's pattern.
+     *
+     * @param member the field or method to check
+     * @return specified pattern (or null)
+     * @see Pattern
+     */
+    protected String resolveStringPattern(MemberScope<?, ?> member) {
+        if (member.getType().isInstanceOf(CharSequence.class)) {
+            Pattern patternAnnotation = this.getAnnotationFromFieldOrGetter(member, Pattern.class);
+            if (patternAnnotation != null) {
+                // @Pattern annotation was found, return its (mandatory) regular expression
+                return patternAnnotation.regexp();
+            }
+            Email emailAnnotation = this.getAnnotationFromFieldOrGetter(member, Email.class);
+            if (emailAnnotation != null && !".*".equals(emailAnnotation.regexp())) {
+                // non-default regular expression on @Email annotation should also be considered
+                return emailAnnotation.regexp();
             }
         }
         return null;